[Librarian] Regenerated @ 125dc871b4170bce0faca169f8a588f6ff356a88 0c222649d98edbac6858ea253f9b77fc22386e47

+++ src/Twilio/Rest/Chat/V2/Service/Channel/InviteResource.cs
@@ -322,7 +322,7 @@ namespace Twilio.Rest.Chat.V2.Service.Channel
                                                      string pathServiceSid,
                                                      string pathChannelSid,
                                                      List<string> identity = null,
-                                                     int? pageSize = null,
+                                                     long? pageSize = null,
                                                      long? limit = null,
                                                     ITwilioRestClient client = null)
         {
@@ -343,7 +343,7 @@ namespace Twilio.Rest.Chat.V2.Service.Channel
                                                                                              string pathServiceSid,
                                                                                              string pathChannelSid,
                                                                                              List<string> identity = null,
-                                                                                             int? pageSize = null,
+                                                                                             long? pageSize = null,
                                                                                              long? limit = null,
                                                                                             ITwilioRestClient client = null)
         {
