+++ src/Twilio/Rest/IpMessaging/V1/Service/Channel/MemberResource.cs
@@ -322,7 +322,7 @@ namespace Twilio.Rest.IpMessaging.V1.Service.Channel
                                                      string pathServiceSid,
                                                      string pathChannelSid,
                                                      List<string> identity = null,
-                                                     int? pageSize = null,
+                                                     long? pageSize = null,
                                                      long? limit = null,
                                                     ITwilioRestClient client = null)
         {
@@ -343,7 +343,7 @@ namespace Twilio.Rest.IpMessaging.V1.Service.Channel
                                                                                              string pathServiceSid,
                                                                                              string pathChannelSid,
                                                                                              List<string> identity = null,
-                                                                                             int? pageSize = null,
+                                                                                             long? pageSize = null,
                                                                                              long? limit = null,
                                                                                             ITwilioRestClient client = null)
         {
